Modify Components and Pages
