+++ src/components/OrderFeed/OrderFeed.tsx
@@ -3,8 +3,8 @@ import styles from './OrderFeed.module.css';
 import { v4 as uuidv4 } from 'uuid';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from '../../services/hooks/hooks';
-import OrderFeedItem from '../../components/OrderFeedItem/OrderFeedItem';
-import OrderFeedStat from '../../components/OrderFeedStat/OrderFeedStat';
+import OrderFeedItem from '../OrderFeedItem/OrderFeedItem';
+import OrderFeedStat from '../OrderFeedStat/OrderFeedStat';
 import { WS_CONNECTION_CLOSED, WS_CONNECTION_START } from '../../services/actions/WSActions';
 import { selectUserCredentials } from '../../services/reducers/authReducer';
 import { Button } from '@ya.praktikum/react-developer-burger-ui-components';
@@ -27,8 +27,8 @@ export const OrderFeed: FC = () => {
     };
 
     useEffect(() => {
-        isLoading(true);
         dispatch({ type: WS_CONNECTION_START, payload: '/all' });
+        isLoading(true);
         return () => { dispatch({ type: WS_CONNECTION_CLOSED }); }
     }, [dispatch]);
 
@@ -38,7 +38,6 @@ export const OrderFeed: FC = () => {
             <div className={styles.wrapper}>
 
                 <div className={`${styles.w100} ${styles.scrollable}`}>
-
                     {loading
                         ? orders.map((order) => (
                             <OrderFeedItem
@@ -53,7 +52,7 @@ export const OrderFeed: FC = () => {
                         : <div>
                             <div className={`${styles.orderFeedEmptyItem} mb-8`}>
                                 <h2 className='mb-4'>Заказов нет!</h2>
-                                <p>Но вы всегда можете создать новый!</p>
+                                <p className='text text_type_main-default'>Но вы всегда можете создать новый!</p>
                             </div>
                             <div className={`${styles.orderFeedEmptyButton}`}>
                                 {AuthUser ? (
@@ -81,7 +80,10 @@ export const OrderFeed: FC = () => {
                 </div>
 
                 <div className={styles.w100}>
-                    <OrderFeedStat />
+                    {loading
+                        ? <OrderFeedStat />
+                        : <OrderFeedStat isEmpty />
+                    }
                 </div>
 
             </div>
