+++ src/pages/ProfilePage/ProfilePage.module.css
@@ -2,51 +2,6 @@
     height: 80vh;
     display: flex;
     align-items: center;
-    /* justify-content: space-between; */
     margin: 0 auto;
     max-width: 1280px;
-}
-
-.content {
-    margin: auto;
-    text-align: center;
-}
-
-.link {
-    color: inherit;
-    /* color: var(--color-title); */
-}
-
-.container {
-    max-width: 320px;
-}
-
-.listLinks {
-    display: flex;
-    flex-direction: column;
-}
-
-.listLinks li {
-    display: flex;
-    align-items: center;
-    height: 64px;
-    width: 320px;
-}
-
-.listLinks li:hover {
-    cursor: pointer;
-    color: #fff;
-}
-
-.activeNavLink {
-    color: white;
-}
-
-.contentHeight {
-    height: 340px;
-}
-
-.text_align_buttons {
-    display: flex;
-    justify-content: right;
 }
